requirements.txt
torch
torchvision
tqdm
matplotlib
opencv-python
sewar
pillow
numpy
scipy
tensorboardx